requirements.txt
# Core dependencies
requests>=2.31.0
httpx>=0.27.0
typesense>=0.21.0
pandas>=2.1.0
pyyaml>=6.0
python-dotenv>=1.0.0
tenacity>=8.2.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# API Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# MCP (Model Context Protocol)
mcp>=1.0.0

# Development dependencies (optional)
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
black>=24.1.0
ruff>=0.1.0
mypy>=1.8.0
types-requests>=2.31.0
